feat: Ignore case enum parsing
Issue scenario:
We are serializing our own object (CompanyModel) which contains enum (Pending) with camel-case resolver. The serilalized json string contains now the enum string camel-cased (pending). Now I try to deserialize this json string back to CompanyModel object but I get an error that says that 'pending' enum is missing, but it is not missing it just written as 'Pending' in the Model. 

We need to ignore case  for enum parsing.
Can you approve this PR?